Add some initial components of C++ foundation library
The foundation library provides common classes and utilities for all Chord C++ projects. This will eventually include the engine, a stand-alone runtime for testing, the VST plugin, and all native module DLLs. The foundation library uses modern C++ modules. This commit adds the following foundation library components:
- Core:
  - Core: should be imported into all files, includes core utilities
  - Types: defines standard type aliases such as s32 and f64 which should be used throughout projects
  - Concepts: common set of convenient concepts
  - Assert: provides assertion mechanisms (this is currently stubbed in)
  - NullTerminatedString: provides basic operations on null-terminated char* and char32_t strings
  - Utilities: various common core definitions and utilities
- Containers:
  - Span: represents a typed span of memory using a pointer and count
  - FixedArray: an array with a fixed element count
  - ResizableArrayBase: base class for resizable arrays
  - BoundedArray: resizable array which can be resized up to a fixed capacity
  - UnboundedArray: resizable array that can be resized in an unbounded manner as necessary
- String:
  - String: string class for ASCII and unicode (UTF32) string operations
- Utilities:
  - BitOperations: utilities that have to do with bit manipulation
  - Bounds: functions such as min, max, etc.
  - Copy: functions to copy ranges of memory

Unit tests have not yet been written for any of the foundation components.

+++ Compiler/Native/NativeTypes/NativeTypes.cs
@@ -56,7 +56,7 @@ internal unsafe struct DataType
 internal unsafe struct NativeModuleParameter
 {
   public ModuleParameterDirection Direction;
-  public sbyte* Name;
+  public uint* Name;
   public DataType DataType;
   public NativeBool DisallowBufferSharing;
 }
@@ -64,7 +64,7 @@ internal unsafe struct NativeModuleParameter
 [StructLayout(LayoutKind.Sequential)]
 internal unsafe struct NativeModuleSignature
 {
-  public sbyte* Name;
+  public uint* Name;
   public NativeModuleParameter* Parameters;
   public SizeT ParameterCount;
   public int ReturnParameterIndex;
@@ -310,7 +310,7 @@ internal unsafe struct ConstantOptimizationRuleComponentValue
   [FieldOffset(0)] public double DoubleValue;
   [FieldOffset(0)] public int IntValue;
   [FieldOffset(0)] public NativeBool BoolValue;
-  [FieldOffset(0)] public sbyte* StringValue;
+  [FieldOffset(0)] public uint* StringValue;
 }
 
 [StructLayout(LayoutKind.Sequential)]
@@ -358,7 +358,7 @@ internal unsafe struct OptimizationRuleComponent
 [StructLayout(LayoutKind.Sequential)]
 internal unsafe struct OptimizationRule
 {
-  public sbyte* Name;
+  public uint* Name;
   public OptimizationRuleComponent* InputPattern;
   public OptimizationRuleComponent** OutputPatterns;
 }
@@ -376,7 +376,7 @@ internal unsafe struct NativeLibrary
 {
   public fixed byte Id[16];
   public NativeLibraryVersion Version;
-  public sbyte* Name;
+  public uint* Name;
   public delegate* unmanaged[Cdecl]<void*> Initialize;
   public delegate* unmanaged[Cdecl]<void*, void> Deinitialize;
   public delegate* unmanaged[Cdecl]<void*, void*> InitializeVoice;
